backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
channels==3.0.2
django-filter==2.4.0
djangochannelsrestframework==0.1.1
drf-yasg==1.20.0
pytest==6.1.2
pytest-django==4.1.0
pytest-asyncio==0.14.0
channels_redis==3.2.0
Pillow==8.0.0
openpyxl==3.0.3
requests==2.24.0
bs4==0.0.1
django-image-cropping==1.5.0
easy-thumbnails==2.7
django-cors-headers==3.5.0